backend/pyproject.toml
[tool.poetry]
name = "desafio-gslab-backend"
version = "0.1.0"
description = "Desafio GSlab"
authors = ["Pbezerra-dev <herbetyp@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.2"
Django = "^3.1.6"
python-decouple = "^3.4"
psycopg2-binary = "^2.8.6"
dj-database-url = "^0.5.0"
django-widget-tweaks = "^1.4.8"
django-cors-headers = "^3.7.0"
djangorestframework-simplejwt = "^4.6.0"
itsdangerous = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"
black = "^20.8b1"
pre-commit = "^2.10.0"
model-bakery = "^1.2.1"
django-extensions = "^3.1.0"
ipython = "^7.20.0"
jedi = "0.17.2"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
